refactor: rename commands::repl to commands::console
